% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data_to_shp.R
\name{add_data_to_shp}
\alias{add_data_to_shp}
\title{Add election results to department geometry}
\usage{
add_data_to_shp(year, round)
}
\arguments{
\item{year}{The election year in the form of YYYY.}

\item{round}{The election round (1 or 2).}
}
\value{
A \code{SpatialPolygonsDataFrame} with values and labels to plot/map.
}
\description{
Add election results to department geometry
}
\examples{
\dontrun{
add_data_to_shp(2022, round = 1)
}
}
