% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_html_content.R
\name{create_html_content}
\alias{create_html_content}
\title{Create HTML page (iframe) with all maps, barplots, and piecharts}
\usage{
create_html_content(year)
}
\arguments{
\item{year}{The election year in the form of YYYY.}
}
\value{
No return values, writes HTML content in \code{content/core.html}.
}
\description{
Create HTML page (iframe) with all maps, barplots, and piecharts
}
\examples{
\dontrun{
create_html_content(2022)
}
}
