% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_election_results.R
\name{get_election_results}
\alias{get_election_results}
\title{Download election results from Minister of Interior}
\usage{
get_election_results(year)
}
\arguments{
\item{year}{The election year in the form of YYYY.}
}
\value{
No return value.
}
\description{
Works well for first round...
}
\examples{
\dontrun{
get_election_results(2022)
}
}
