% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tooltip_barplot.R
\name{make_tooltip_barplot}
\alias{make_tooltip_barplot}
\title{Create HTML Barplots (Tooltip) pour first round of voting}
\usage{
make_tooltip_barplot(year)
}
\arguments{
\item{year}{The election year in the form of YYYY.}
}
\value{
No return values. Barplots (HTML code) are exported in
\verb{content/barplots/}.
}
\description{
Create HTML Barplots (Tooltip) pour first round of voting
}
\examples{
\dontrun{
make_tooltip_barplot(2022)
}
}
