% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tooltip_piechart.R
\name{make_tooltip_piechart}
\alias{make_tooltip_piechart}
\title{Create HTML Piecharts (Tooltip) pour second round of voting}
\usage{
make_tooltip_piechart(year)
}
\arguments{
\item{year}{The election year in the form of YYYY.}
}
\value{
No return values. Piecharts (HTML code) are exported in
\verb{content/piecharts/}.
}
\description{
Create HTML Piecharts (Tooltip) pour second round of voting
}
\examples{
\dontrun{
make_tooltip_piechart(2022)
}
}
