% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_candidates_info.R
\name{read_candidates_info}
\alias{read_candidates_info}
\title{Read candidates information}
\usage{
read_candidates_info(year)
}
\arguments{
\item{year}{The election year in the form of YYYY.}
}
\value{
A \code{data.frame} with three columns:
\itemize{
\item \code{name}: the name of the candidate (for joins);
\item \code{label}: the name of the candidate (for display);
\item \code{color}: the color associated to the candidate.
}
}
\description{
Imports a CSV file stored in \verb{data/YYYY/} and named
\code{candidates_info_YYYY.csv} where \code{YYYY} is the year of the election.
}
\examples{
\dontrun{
info <- read_candidates_info(2022)
}
}
