% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_election_results.R
\name{read_election_results}
\alias{read_election_results}
\title{Read election results}
\usage{
read_election_results(year, round)
}
\arguments{
\item{year}{The election year in the form of YYYY.}

\item{round}{The election round (1 or 2).}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item \code{no_departement}: the number of the department;
\item \code{departement}: the name of the department;
\item \code{inscrits}: the number of registered electors;
\item \code{abstention}: the number of abstentions;
\item \code{votants}: the number of voters;
\item \code{blancs}: the number of blank votes;
\item \code{nuls}: the number of spoiled votes;
\item \code{exprimes}: the number of expressed votes;
\item \code{xxx}: the number of votes for the candidate xxx.
}
}
\description{
Imports an XLS file stored in \verb{data/YYYY/} and named
\code{Presidentielle_YYYY_Resultats_Tour_X_c.xls} where \code{YYYY} is the year of the
election and \code{X} is the round.
}
\examples{
\dontrun{
votes <- read_election_results(2022, round = 1)
}
}
