% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_alone.R
\name{theme_alone}
\alias{theme_alone}
\title{Alone theme for highcharts}
\usage{
theme_alone(...)
}
\arguments{
\item{...}{A named parameters to modify the theme.}
}
\description{
Derived from version 0.9.0 of Highcharter
}
\examples{
\dontrun{
highcharter::highcharts_demo() \%>\%
  highcharter::hc_add_theme(theme_alone())
}
}
